requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
Django==5.1.6
django-taggit==5.0.1
gunicorn==23.0.0
h11==0.14.0
Markdown==3.6
packaging==24.2
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
uvicorn==0.34.0
whitenoise==6.9.0
